requirements.txt
django==4.1
djangorestframework>=3.13.1,<3.14
django-cors-headers>=3.13.0,<3.14
requests>=2.28.1,<requests==2.29